Revert "fee to double"
This reverts commit c78f976388cd990056d07cdc98ebb9a7df807f42.

+++ icdc_android_app/lib/resources/patient_object.dart
@@ -15,5 +15,5 @@ class DentalRecord {
   String surface = "";
   String description = "";
   int transDate = 0;
-  double fee = 0;
+  int fee = 0;
 }
